requirements.txt
Flask==3.1.1
requests==2.32.4
openai==1.86.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2

